Project41/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.22.0",
    "asyncpg>=0.30.0",
    "flask>=3.1.2",
    "pillow>=11.3.0",
    "python-telegram-bot>=22.5",
    "requests>=2.32.5",
    "telegram>=0.0.1",
    "tlgbotfwk>=0.4.61",
]
